refactor: 모듈 파일 네이밍 변경
[ModuleName]/module.js 에서 [ModuleName].js의 형태로 변경
