+++ src/domain/RacingWinnerRecorder/RacingWinnerRecorder.js
